+++ test/jenkins/build.sh
@@ -30,10 +30,6 @@ case "$(hostname)" in
         echo See compile.log for more information!
         ;;
 
-    kesch*)
-        echo Extpar is no longer supported on Kesch!
-        ;;
-
     tsa*)
         run_command git submodule init
         run_command git submodule update
@@ -47,23 +43,19 @@ case "$(hostname)" in
         ;;
 
     # DKRZ machines    
-    mlogin*)
-        if [[ -r /sw/rhel6-x64/etc/profile.mistral ]]
+    *levante*)
+        if [[ -r /sw/etc/profile.levante ]]
         then
-           source /sw/rhel6-x64/etc/profile.mistral
+           source /sw/etc/profile.levante
         fi
-        run_command source /etc/profile.d/mistral.sh
         run_command git submodule init
         run_command git submodule update
-        run_command ./configure.mistral.$compiler
+        run_command ./configure.levante.$compiler
         run_command source modules.env
         run_command make clean
         echo compile extpar...
         run_command make &> compile.log
         echo          ...done
         echo See compile.log for more information!
-
         ;;
 esac 
-
-