+++ test/testsuite/data/get_data.sh
@@ -60,39 +60,4 @@ wget --quiet 'ftp://iacftp.ethz.ch/pub_read/juckerj/corine/external_parameter_ic
 wget --quiet 'ftp://iacftp.ethz.ch/pub_read/juckerj/corine/clim_tsea_corine.nc'
 wget --quiet 'ftp://iacftp.ethz.ch/pub_read/juckerj/corine/clim_t2m_corine.nc'
 cd -
-
-# intel references
-cd intel
-
-# mch
-test -d mch || exit 1
-cd mch/c7_globe
-wget --quiet 'ftp://iacftp.ethz.ch/pub_read/juckerj/external_parameter_intel_mch_c7_PR207.nc'
-cd -
-
-test -d mch || exit 1
-cd mch/c1_aster
-wget --quiet 'ftp://iacftp.ethz.ch/pub_read/juckerj/external_parameter_intel_mch_c1_PR164.nc'
-cd -
-
-# clm
-test -d clm || exit 1
-cd clm/12km_globe
-wget --quiet 'ftp://iacftp.ethz.ch/pub_read/juckerj/external_parameter_intel_12km_globe_PR278.nc'
-cd -
-
-# dwd
-test -d dwd || exit 1
-
-cd dwd/icon_d2
-wget --quiet 'ftp://iacftp.ethz.ch/pub_read/juckerj/external_parameter_intel_icon_d2_PR273.nc'
-cd -
-
-# mpim
-test -d mpim || exit 1
-cd mpim/icon_r2b4
-wget --quiet 'ftp://iacftp.ethz.ch/pub_read/juckerj/external_parameter_intel_icon_mpim_PR286.nc'
-cd -
-cd ..
-
 # done