+++ test/testsuite/submit.levante.sh
@@ -5,7 +5,7 @@
 #SBATCH --cpus-per-task=12 
 #SBATCH --output="job.out"
 #SBATCH --time=02:00:00
-#SBATCH --partition=gpu,prepost,shared,compute,compute2,
+#SBATCH --partition=gpu,shared,compute
 #SBATCH --account=mh0287
 
 export OMP_NUM_THREADS=$SLURM_CPUS_PER_TASK